Add files via upload
Triple Threat

Triple Threat is an engaging console-based game that challenges players to manage randomly generated arrays containing five elements. The primary objective is to forming or encountering a sequence of three consecutive identical elements. Once such a sequence is detected, the game ends, making it both a test of strategy and luck.

Gameplay Mechanics:

Initialization:

At the start of the game, an array consisting of five elements is generated randomly. These elements are selected from a predefined set of values, which can include numbers, letters, or symbols. This ensures that each game session begins with a unique configuration, adding to the unpredictability and excitement.

Display:

The initial array is displayed on the console for the player to view. This allows the player to analyze the arrangement and plan their next moves accordingly.

Input:

Players are given the option to play the game by pressing a specific keyword and to end the game by entering zero. These actions are crucial as the player will never know what value will be present next.

Detection:

After every player action, the game checks the array for any sequence of three consecutive identical elements. This is a interconnected  part of the game logic as it determines whether the game continues or ends based on the player's moves.

Game Over:

If the game detects three consecutive identical elements in the array, it immediately ends with Winning message ,otherwise the player is notified with a Game Over type of message. This instant feedback loop ensures that players are aware of the consequences of their actions.